net: vdsClient fails to connect to an IPv6 addr
When vdsClient is used on a non-ipv6  host, and no destination address
is provided, it will fail to connect to the server.
(it uses the vdsm.conf management_ip parameter which is by default '::')

The vdsClient now prompts when it detects/suspects this is the problem.
(instead of a traceback)

Change-Id: I069dcb3863a720f414d3048299fd54ef74839bdb
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/52367
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Continuous-Integration: Jenkins CI