lun: Serial attr should not passed to libvirt for lun disks.
Following the change which was introduced at
libvirt-1.2.17-5 and qemu-kvm-rhev-2.3.0-18, libvirt now enforce
that a scsi-block 'lun' device should not provide the 'serial' property.

To comply with this change, this patch introduces a change
that a VM should ignore 'serial' property for lun devices when
passing the XML to libvirt.

Since the serial attribute is basically redundant for lun devices
it was also removed for EL6 hosts which are used in oVirt 3.5

This patch was verified by running a VM with lun disk which
has enabled SCSI Pass-Through and also plugging the same lun disk
while the VM was running.
It was also verified on EL6 running libvirt-0.10.2-54.el6_7.2.x86_64.
The operation succeeded.

Change-Id: I41abb2b7d7f5950e65710379a0655399014afd83
Bug-Url: https://bugzilla.redhat.com/1299235
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>
Tested-by: Maor Lipchuk <mlipchuk@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/51766
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/52463
Continuous-Integration: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Francesco Romani <fromani@redhat.com>