periodic: ignore known-benign libvirt errors
periodic sampling threads and migration thread are
(and need to be) asynchronous with each other.
This leaves the door open for errors like

libvirtError: Domain not found: no domain with matching uuid
'7ebb5925-e76f-4d2f-a148-7671c313fe84' (v2)

Which are harmless but spam the logs.
We don't want to hide real errors, but we want to
avoid this noise, so this patch makes periodic Operation
ignore some of these errors, the ones which are easier
to detect.

Change-Id: Ifb5618f44870bdca97fa479734514a3c8d81c7d7
Bug-Url: https://bugzilla.redhat.com/1299480
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/52884
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53999