Live Merge: Skip initial extension when base is raw
When starting a live merge we pre-extend the base volume in order to
accomodate the data that will be merged into the base volume.  This is
only necessary for thinly provisioned block volumes.  Check to make sure
that the base volume is eligible for extension before extending it.

Change-Id: If105854174e9fdbc9bad7aef511ff233de8991a3
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1308626
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53110
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Ala Hino <ahino@redhat.com>
Continuous-Integration: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>