vmfex hook: hold a ref for a virDomain object
commit 93feb1c introduced a weakref to virDomain object to avoid
reference cycles. This causes a surprising behaviour: if we do not hold
a reference to virdomain.Notifying, the underlying virDomain is
garbage-collected before being accessed for the first time.

This patch quickly holds a named reference to the domain. I did not find
other pieces of code with the same problem, however, the mechanism seems
dangerously error-prone.

This bug is easily reproduced with

    from vdsm import libvirtconnection
    c = libvirtconnection.get()
    print(c.lookupByName('my-vm').XMLDesc(0))

and "solved" by

    from vdsm import libvirtconnection
    c = libvirtconnection.get()
    v = c.lookupByName('my-vm')
    print(v.XMLDesc(0))

Change-Id: Ida7488db5f8af51da5a2931d8b8a6c51d99c36f0
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1305338
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53160
Reviewed-on: https://gerrit.ovirt.org/53203
Reviewed-by: Francesco Romani <fromani@redhat.com>