virt: Don't use Vm device configuration in clientIF
clientIF retrieves device configuration from Vm.devSpecMapFromConf() in
order to pass it to Vm.preparePaths().  This is unnecessary exposition
of device handling configuration to clientIF as Vm can get the
information from itself.  Not only it's not clientIF business to care
about those details, we'd also like to make them isolated in Vm and make
Vm.devSpecMapFromConf() private in a followup patch.

This is a preparation step for moving device configuration processing
from Vm to device classes.

For those reasons we retrieve configuration in Vm.preparePaths() from
its instance directly.

Change-Id: Idd0f7c99b4df18d27dfad36bf2275540bc8b958c
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53482
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>