migration: Add Dynamic Bounded Semaphores
This patch adds new Bounded Semaphore class with the ability to modify the
bound at runtime. It also migrates the existing `incoming` and `outgoing`
migration semaphores to use this new dyanmic version.

This is useful for subsequent migration enhancements that will allow the engine
to dynamically adjust the maximum number of concurrent migrations.

Change-Id: I67376d0bd990f89e0a013887cef1a0cb05fb855d
Signed-off-by: Martin Betak <mbetak@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53304
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>