schema: inconsistency fixes
There was number of inconsistencies in schema where we specify in
description that a parameter is optional whereas in type definition is
not the case. To ease conversion of the schema to yaml format
we need to make sure that inconsistencies are fix.

I made assumptions that description is correct and fixed type
definition.

Change-Id: I95f4f639f3d7de1179f4c77b5267134e88044fd3
Signed-off-by: pkliczewski <piotr.kliczewski@gmail.com>
Reviewed-on: https://gerrit.ovirt.org/53460
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>