virt: Pass effective guest agent API version during migration
Currently we're not passing the effective guest agent API version
during migrations, since VDSM and the guest agent can negotiate this
version.

For the feature of lifecycle events for the guest it is needed that
VDSM already knows the negotiated API version right from the moment
of the migration completion so we can notify the guest about the
completion of the migration.

This will save VDSM from having to re-negotiate the API version,
and allow us to send the guest agent the required information right
away.

In case the passed version is higher than supported, the loweest
common maximum version will be chosen which will require
negotiation later on, however it will give VDSM the ability to
send the event anyway since it implies this feature to be present
on both VDSM sides of the migration.

Change-Id: I0e8625ad33f62852baa17cad377ab12065a52dec
Bug-Url: https://bugzilla.redhat.com/1298120
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53523
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI