virt: Move Vm._getUnderlyingDeviceAddress() to vmxml.py
This is a preparation step for the next round of moving device handling
code out of Vm class.  We want to move Vm._getUnderlying* methods out of
Vm, but not all at once.  Most of the methods use
_getUnderlyingDeviceAddress() so we need to have this functionality
available in both Vm and the moved code and this is the reason for
having this preparation step.

Change-Id: Ib9868ca1df6650b1262e6f30d639a08b1f38304d
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53615
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI