virt: Move Vm._getUnderlyingSoundDeviceInfo() out of Vm
This is the first sample movement of _getUnderlying*DeviceInfo() method
movement out of Vm class.  This patch does nothing more than moving the
corresponding code and updating the docstring accordingly.  There are
other tempting changes we might want to do with the code, such as
getting rid of the configuration updates, but that's better done later,
after everything is moved to proper places.

Other _getUnderlying*DeviceInfo() methods are going to be moved in the
same way in the followup patches.

Change-Id: I05cb24e0a9776e21bd2546d62a56554d74559c61
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53616
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>