virt: periodic: disable on migration destination
While a VM is migration destination, the periodic
operations should not run - the VM is still on source side,
and the source side should handle the housekeeping.

We should start doing the operations on destination side
when and if the VM succesfully migrated.

Previously, we didn't made this explicit out of concern
of hiding real bugs.
This concern is more and more less grounded, while log
spam (all false positives) is becoming a problem, so
we have reached a tipping point. Hence this patch.

Change-Id: Id27d3524e1d183e9db4e6bfe55873900b41fd4c7
Bug-Url: https://bugzilla.redhat.com/1299480
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53539
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54400