net: Fix internal setupNetworks api usage
setupNetworks api defines 2 args and an unlimited number of kwargs.
The code used in the global api to proxy the request to supervdsm used
3 args (the last being a dict).

To be consistent and avoid multipe conversions, setupNetworks definition
should specify 3 args.

This patch is a preperation for removing the bridging that exist at
supervdsm api and exporting setupNetworks api dynamically.

Change-Id: I96e2b7be3251a1c20111034e4cc6a4b746dde269
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53572
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Tested-by: Dan Kenigsberg <danken@redhat.com>