fileUtils: Add atomic_symlink utility
This helper function creates or updates a symlink atomically. If the
symlink exists but links to a different target, it is replaced
atomically with a link to the specified target.

We use this logic now for linking storage domain, and we want to use the
same logic for creating or updating the image rundir links
(/run/vdsm/storage/sd_uuid/img_uuid)

Change-Id: I931eca7a25c7b10344fce78b274d633d73eb082d
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53680
Continuous-Integration: Jenkins CI
Reviewed-by: Adam Litke <alitke@redhat.com>