virt: recovery: test: fix typo
A typo sneaked in the vdsm.virt.recovery.File.save() method.
It passed by because this code path is seldom exercised, because
is meant to catch a programming error[1].

This patch fixes the typo, and adds a test to prevent it to
happen again.

+++

[1] in a sneaky way, mostly due to backward compatibility
with surrounding code.
The old 3.6.x code just swallowed any error.
Now we have at least a log line.
A better fix would be to explode with exception,
but due to arcane intertwined flows, this may cause havoc.

Change-Id: Ifc0a58758e2878d3c5398e931982d6e193fcd9f1
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53689
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>