net: also disable IPv6 on bridges not using it, after unified restoration
This is a follow-up work to disallow VM-host communication over IPv6
(using e.g. link-local addresses) in the case when networking on the host
was configured without IPv6.

This covers the reboot scenario, in addition to the previously implemented
disablement of IPv6 during network configuration through setupNetworks.

Change-Id: I12be9d8c99ddfeb85d6ba6093535e35b6883e963
Bug-Url: https://bugzilla.redhat.com/1219363
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53683
Continuous-Integration: Jenkins CI
Reviewed-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>