gluster: Don't fail connect server when getting volume info
When connecting to gluster server, we check volume info in order
to validate replica count. However, if gluster server is not
available, we get an exception and fail the operation. This patch
catches the thrown exception and logs a warning message indicating
that we got error when tried to connect to gluster server.

Change-Id: I0e1835b7de4c5c4c5c4616d3c36f15c1f91a01dc
Bug-Url: https://bugzilla.redhat.com/1303977
Signed-off-by: Ala Hino <ahino@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53785
Continuous-Integration: Jenkins CI
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Sahina Bose <sabose@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54689
Reviewed-by: Francesco Romani <fromani@redhat.com>