virt: start moving under site-packages
Let's start moving virt code under site-packages.
We start with the leaf node of the dependency tree,
the virt-specific utility modules.

As nice side effect, these modules are mostly
stable and seldom changes, so once we get the
import and build system right, we are done with the move.

Change-Id: I8628666a97a87e194c2f85d3be03934d0f7fa1c3
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53958
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>