virt: extract destroyVm helper
It is generally not safe to force the destruction
(really: shutdown) of a VM. There are few known cases
on which we believe it is safe despite a previous
failure of destroyVmGraceful().

This patch adds an utility method to encapsulate
force-shutdown-afeter-known-safe-error logic,
previously found in destroyVmGraceful().

No changes in external behaviour.

Change-Id: Id152a7aae3d4f644f34703f7f0b5f98035e4e295
Bug-Url: https://bugzilla.redhat.com/1309884
Backport-To: 3.6
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/53932
Reviewed-by: Dan Kenigsberg <danken@redhat.com>
Continuous-Integration: Jenkins CI