net: api: fix invalid refactoring
else branch used to iterate 'nics' and if there was none, it just did
nothing.

During the refactoring [1] we get rid of 'nics' and replaced it with
'nic'. In case when 'bonding' is None, we go to else branch and
without any checking call _validateInterNetworkCompatibility()
which fails for None nic.

[1] https://gerrit.ovirt.org/#/c/53475/15/lib/vdsm/network/api.py

Change-Id: I7f5ecd7682b05da4970930656b4f18c0f67b4180
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54016
Continuous-Integration: Jenkins CI
Reviewed-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>