response: Add support for specific error checking in is_error
response.is_error can now take an optional 'err' argument to check
if given response is of that particular error type.

E.g. to check if given response if an instance of 'migrateLimit' error:

  response.is_error(result, 'migrateLimit')

This will return false for successful response as well for all error codes
other than that of 'migrateLimit'.

Change-Id: I00d8d2914983f75293ece5ce05894ca3821b7b24
Signed-off-by: Martin Betak <mbetak@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54036
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>