net: ifcfg: rollback nicless bridge
Ifcfg rollback is not able to rollback simple nicless bridge. We do
the removal of interfaces just with ifdown during the rollback. For
some reason this is enough for vlaned networks, but not for
a just-a-bridge network.

We fixed that with explicit delbr of to-be-rolled-back bridges.

Because of we don't guarantee handling order in networks setup, this
cannot be simply tested - we cannot create a test where nicless
network is set up and then another network setup fails.

This patch was tested manually. In a following patch, explicit removal
of networks will be removed from _check_connectivity() and therefore
testSetupNetworksConnectivityCheck() will work as a test for this
scenario.

Change-Id: Ic0cd62806fe10e288cc6183a554b52d97883bc82
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54069
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>