net: rename destroyOnMasterRemoval
destroyOnMasterRemoval parameter was needed to say whether bonding
should be removed together with network (implicit bonding) or just
detached. Now when we drop implicit bonding, we can change it's logic
to something more understandable.

With this patch, destroyOnMasterRemoval was renamed to
on_removal_just_detach_from_network. It's extremely long, but verbosity
is needed here.

Default functionality of bond.remove() is evidently removal, so
this new parameter is False by default.

When a bonding is objectivized as a part of a network, we need it
to be just detached from master, set it default MTU etc, so it
will be a standalone bonding again. When we objectivize it as a
part of bonding query, then by bond.remove() we really mean the
removal.

Change-Id: I0ca4ba568c3f24b20a5e9ee70b3f6935bbc22eab
Signed-off-by: Petr Horáček <phoracek@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54082
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>