repoplot: Add storage domains monitors commands
The time to read from storage is already reported in the repostat
"delay" graph. This graphs show the time to run the dd command. We have
seen cases where the delay was very fast, but running dd command took 15
seconds. I suspect that this issue was caused by overloaded vdsm on
machine with many cores, and should be fixed by using cpu_affinity.

Change-Id: I48b84bf4b9ef7518bbf08675f3dd1a0557c524f9
Relates-To: https://bugzilla.redhat.com/1081962
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54107
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI