repoplot: Avoid creation of empty command stats
Previously we did:

    cs = commands[log.thread]

Which creates a new CommandStats for the current thread. However if the
log did not match the start or end conditions, and no other command was
issued for this thread, this would create a useless record.

Now we access the command only if the start or end conditions match.

Change-Id: Ia76be19c929c4e629bdd335713c7d0184dcdd9b8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54111
Reviewed-by: Francesco Romani <fromani@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI