network: inherit DUID from a bridge's port with no restriction
A bridge must reuse the DHCP unique identifier of the underlying
network device it is built atop so DHCP client obtains the same
lease (in particular, the same address).

Recently, the flow in setupNetworks changed in a way that the device
being "replaced" by the bridge no longer reports 'dhcpv4' as True.
There is simply no evidence of DHCPv4 being used on the device, and
thus DUID would not be inherited, as the checking is too strict.

Now, dhclient run on the bridge is always called with the '-df' option
(used to refer to the bridge port's lease file) as long as it is
supported. The option is harmless: dhclient simply ignores it if the
lease file doesn't exist.

As dhclient doesn't have a -h/--help option, we pass to it an
intentionally invalid option, tricking it into listing the options it
supports. To avoid printing the error output to logs, dhclient is now
called directly, not through execCmd.

Change-Id: Id86f3afc20562ca670370b2e2907d46ae2203900
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54130
Continuous-Integration: Jenkins CI
Reviewed-by: Edward Haas <edwardh@redhat.com>
Tested-by: Edward Haas <edwardh@redhat.com>
Tested-by: Paul Maidment <paulmaidmentuk@googlemail.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>