caps/machinetype: move machine-type related code to new module
Machine types are what qemu considers the computer. Libvirt exposes
them in capabilities XML - the machine types themself and inherently
related information such as baseline CPUs for the machine type.

This patch helps machine type related code to find a home in a new
module called machinetype. The reason for this is that 'caps' module
became code bin over time without single objective. Additional benefit
is that, although currently not done so, the code is exposed and may be
used in hooks.

Change-Id: Ia7bf79c66a38329672e1fbd3fd41a94058803b58
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54212
Reviewed-by: Francesco Romani <fromani@redhat.com>
Continuous-Integration: Jenkins CI