vmstats: avoid NameError
vmstats._nic_traffic may raise NameError if a couple
of fields are missing. This, however, is definitely
possible (and not-critical) when dealing with libvirt
bulk stats.
Previously, the code was only looking to be safe against
KeyError.

Now we make sure to not use variables which may not exist.

Along the way, we improve docstrings and tests for
the related functions in vmstats.py

Change-Id: I35d6db82941f1a8f4c59b23df96b00be73814579
Signed-off-by: Francesco Romani <fromani@redhat.com>
Backport-To: 3.6
Reviewed-on: https://gerrit.ovirt.org/54247
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>