virt: more moving under site-packages
Move another round of virt modules under site-packages.
The modules left out are the one still actively
worked, as well as the dreaded vm.py, as part as the ongoing
effort to cleanup VM devices handling and move from minidom
to etree.
Same applies for the migration code, which is being changed
because of the 'migration enhancements' feature.

We will have another mass move later in the cycle,
when the aforementioned work is done.

Change-Id: Ibddcae2ead4fbc270a5ad021b09a4fa322c534b1
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54366
Continuous-Integration: Jenkins CI
Reviewed-by: Vinzenz Feenstra <vfeenstr@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>