tests: incorporate pool id in file domain path
In real environments, an image 'A' in a file domain would appear in the
following directory path:

    /<storage-repo>/<pool-id>/<domain-id>/images/A

Prior to this patch the helper function make_filesd_manifest was
creating the structure:

    /<storage-repo>/<domain-id>/images/A

This obviously doesn't agree and needs to be fixed.  The solution is to
generate a spuuid in make_filesd_manifest and construct the metadata
appropriately and then initialize the domain manifest with the path that
includes the pool id.

To make sure this works for all callers, prohibit callers from passing
in their own domain metadata.  Only one test needed to be fixed to make
this change.

Change-Id: I1bac433a081b058eec4e1897ea85e698850abb2f
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54565
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>