schema: add Rng and Tpm device to the device list
If Engine asks Vdsm to create a Vm with Rng or Tpm device,
then Vdsm
- consumes the data from the VM.create arguments
- reports the devices in the VM.list (et. al.) return value

The problem here is just the Rng and Tpm device were not listed
as valid device types, so they *seem* unused.

I actually run a VM with rng device enabled and seen that this
data is indeed present, so this patch adds the missing bits
to make the schema consistent.

Change-Id: If5ad65b8714e3dfa7bdc8824461bfae2f9dacae1
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54425
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Tested-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>