vm: devices: handle no alias if balloon model=none
A disabled balloon device is translated to the XML
element

  <memballoon model="none" />

in the domain XML Vdsm send to libvirt. In this case,
libvirt should not give back an alias for this device,
but it does.
On restart, however, that alias is lost.
But Vdsm still expects one alias to be present, so
the recovery fails with

  File "/usr/share/vdsm/virt/vm.py", line 4121, in
_getUnderlyingBalloonDeviceInfo
    alias = x.getElementsByTagName('alias')[0].getAttribute('name')
IndexError: list index out of range

The end result is that the VM is recovered, but reported 'paused',
while is actually running fine.

This patch makes Vdsm handle balloon devices without alias.

Change-Id: Iab96f47118148a67135286a3a9b1998840efce18
Backport-To: 3.6
Bug-Url: https://bugzilla.redhat.com/1335840
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54440
Continuous-Integration: Jenkins CI
Reviewed-by: Milan Zamazal <mzamazal@redhat.com>
Reviewed-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>