ifcfg: re-enable IPv6 before device configuration, or disable afterwards
If a network has been restored with no IPv6 requested, all its devices
have disable_ipv6 set to 1 and up to now there was no other way to
re-enable IPv6 than doing so manually or rebooting.

This patch enables IPv6 on the top-level device ("the network") prior
to ifup, sets the underlying devices' disable_ipv6 to 1 as they are
not directly involved in communication. If IPv6 is not requested, it
is disabled on all devices after ifup.

Change-Id: Idddfb096e6ea384dbe6655c5c4178d4884a8db85
Signed-off-by: Ondřej Svoboda <osvoboda@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54555
Continuous-Integration: Jenkins CI
Reviewed-by: Edward Haas <edwardh@redhat.com>
Tested-by: Edward Haas <edwardh@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>