storagetests: initial FakeEnv support
Many storage tests rely on a mocked storage environment in order to test
the storage code.  We can achieve this environment through a combination
of mocked objects, monkey-patching, and use of temporary files and
directories.  To simplify the tests and increase reliability we
introduce a standardized interface implemented with context managers
that can be used by all tests.  While not all tests actually require the
full scope of these mocked environments, the benefits of uniformity
justify the extra setup.

Change-Id: I23c27ad8f642639852796ab98fb102dedf55aedd
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54567
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI