tests: Provide dnsmasq_run as a context manager
The functional tests are using dnsmasqDhcp as the dnsmasq context
manager. It is limited to a single configuration and located as part of
the functional network test module.

Same functionality has been added in the nettestlib module, with options
to allow a test to specify its own configuration, per the needs.

dnsmasqDhcp is replaced by dnsmasq_run.

Change-Id: Ifde1232416081981256c3e384151574617e9b0d0
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54632
Continuous-Integration: Jenkins CI
Reviewed-by: Petr Horáček <phoracek@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>