tests: Add test_overwrite_blocksize test
The BlockStorageDomainManifest has special behavior with respect to its
logBlkSize and phyBlkSize properties.  These are set at object
construction time from the passed metadata but are not updated when
replaceMetadata is called (even if the new metadata specifies new
sizes).  This test verifies that values are not overridden by
replaceMetadata but that non-default values can be provided during
object construction.

Change-Id: I85fc8ad393aff49154d9d414bd6474a723b315f1
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54669
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-by: Freddy Rolland <frolland@redhat.com>