directio: Add failing test for DirectFile.readlines()
Unlike built-in file object, or io.FileIO, line terminators are not kept
in the output of readlines(). This is usually convenient, but it is not
the expected and documented behavior.

This patch add a new failing test and mark it as brokentest. We will fix
this in the next patch.

Change-Id: I9b487a6537efcf38e61a9170835124cab8821531
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54694
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI