directio: Require full blocks
Previously write accepted a partial block, silently padding the data
with zeros, possibly overwriting data.

For example, if file contents was:

    b"a" * 512 + b"b" * 512

Writing "c" * 768 would change to file to:

    b"c" * 768 + b"\x00" * 256

The correct way to perform such write is to read 2 blocks, modify the
data, and write back 2 blocks. But this is inefficient, so it must be
implemented in the caller of DirectFile. It also does not make sense in
the context of directio.

Now this write will fail with ValueError:

    ValueError: You can only write in 512 multiplies

Change-Id: I6c13e322926c424b83ae09d9d427e71598f99f3b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54715
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI