directio: Remove useless DirectFile.writelines()
Writing lines does not makes sense in the context of directio, since
writes must be aligned to blocks, and there is tiny chance that
arbitrary list of string will come out aligned after joining.

This functionality belongs to higher layer, formating lines of text and
padding the data if needed. And indeed caller of DirectFile already
implement this.

Change-Id: I0c3e03c4d96f1862df2c5fec834cc3787145a3fa
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54716
Reviewed-by: Freddy Rolland <frolland@redhat.com>
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI