virt: common handling of exceptions
Vm methods should not return responses directly.
Rather, they should raise proper exception,
and a common layer should translate them in responses.

In case of succesfull completion, this code automatically
produces a generic success response. Should the method
require extra fields, it can return a dictionary
with the keys to be added to the success response.

Using this way the method could also override the default
keys (e.g. the message). This is not clean but it could
be needed for backward compatibility.

Change-Id: Ic186dc8fa062d8b3789c6057bba68fbbc23f311b
Bug-Url: https://bugzilla.redhat.com/1316128
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54664
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>