monitoring: Introduce the asyncevent module
This module provides a simple asyncore based event loop based on the
asyncio.BaseEventLoop from Python 3. This module will be used for new
storage domain monitoring and for replacing code using bare epoll and
poll.

This module is compatible with both Python 2 and 3.  When we require
Python 3, we can drop this module and simplify client code using asyncio
coroutines.

Although this is a generic infrastructure, keeping it in storage until
it is needed in other parts. The event loop may replace the vmchannels
and yajsonrpc event loops.

Since this file was mostly copied from Python 3, it is licensed under
the python license.

Change-Id: I627f06f136792967f257b6ec439622432f2021be
Bug-Url: https://bugzilla.redhat.com/1081962
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54849
Reviewed-by: Adam Litke <alitke@redhat.com>
Continuous-Integration: Jenkins CI