tests: Run unit tests using travis-ci and docker
Integrate with travis ci to run unit tests with the help of a vdsm
docker image.

Currently it is based on the docker image edwardhaas/ovirt-vdsm-dev that
contains the required vdsm packages for running the unit tests.
The docker image is run, mounting current branch/pull commit and executes
the unit tests (make check-unit).

The image can be also be run locally:
- Install docker.
- Get and run the image: docker run edwardhaas/ovirt-vdsm-dev

Travis CI run example, with this patch, can be reviewed here:
https://travis-ci.org/EdDev/vdsm

Travis CI should be enabled on the ovirt/vdsm repo so it can run on each
merge.

Change-Id: Ia6cd647acb4c740f890e9b602783915439317880
Signed-off-by: Edward Haas <edwardh@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54856
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>