hostdev: don't implicitly reattach devices
We have previously dealt with device drivers as follows:

1) detach (means unbinding from host driver),
2) VM runs and
3) reattach (binds the driver back to host).

This patch removes step 3. Although it explicitly makes device unusable
for the host after VM's runtime, the patch series also re-introduces
the hostdevReattach verb to explicitly fix these devices.

The importance of this change lies in the fact that device drivers are
evil. Rebinding to a driver may lead to horrible issues, most notably
host kernel crashes. We can also assume that, on properly maintained
hosts, the devices are reserved for assignment and are not meant to be
used by the host at all.

Valid points for this kind of behavior are in this[1] thread.

[1] https://www.redhat.com/archives/libvir-list/2016-March/msg00593.html

Change-Id: I1136292cb7e9d13a8202b1f5162299d2155e8036
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54936
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>