hostdev: add support for hotunplug
Hotunplugging host device without implicit reattachment is very similar
to different hotunplugs

Part of an IOMMU group or the whole group can be hotunplugged as there
is no additional consistency check on VDSM's side. As the verb operates
on bulk of devices, the return value informs of which devices were
hotunplugged to allow for keeping track of which were not.

Change-Id: If36f7da8f018615b76a1a8779ff15b890a823db8
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1234283
Signed-off-by: Martin Polednik <mpolednik@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/42747
Continuous-Integration: Jenkins CI
Tested-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Francesco Romani <fromani@redhat.com>