spec: don't depend on ovirt-vmconsole on fedora
Should Vdsm be installed on a system on which the
ovirt-vmconsole package is not installed, it must
be ./configure'd with:

./configure \
  --with-ovirt-vmconsole-user=vdsm
  --with-ovirt-vmconsole-group=kvm

In particular the ovirt-vmconsole-group could be something
else, but MUST be a group already present in the system.
Without those two options, any VM will fail to run.

This is especially relevant for fedora, where the ovirt-vmconsole
package is not yet available.

Change-Id: I81d39abcc55e371ac3631b0d5394ebf174c86fee
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/54999
Continuous-Integration: Jenkins CI
Reviewed-by: Yaniv Bronhaim <ybronhei@redhat.com>
Reviewed-by: Dan Kenigsberg <danken@redhat.com>