lib: set system name for threads
To make the troubleshooting easier, we want
to set meaningful system names for our threads, so
we can use system tools (top, ps) to learn what Vdsm is doing.

The problem is that the system thread name must be at most
15 ASCII character long, while python's thread name could be
any legal string

A good shortening algorithm failed to emerge, so we
just truncate the possibly long python name and we use the truncated
name for system thread name.

Change-Id: I165f5ce9c86dad407d956a3b07e247431f2eed41
Bug-Url: https://bugzilla.redhat.com/1141422
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55074
Reviewed-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>
Reviewed-by: Irit Goihman <igoihman@redhat.com>
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI