fileUtils: tarCopy - avoid extracting the file modified time
We using fileUtils.tarCopy() when there are unsynced clocks between
the source/destination tar might warn us that the source timestamp is
in the future (when we copy from/to file).
This error is logged to stderr, if the copy operation fails for some other
reason the stderr might be flooeded with timestamp in the future warnings
which is confusing when analyzing the cause to the failure.

As we don't care about the timestamp this patch adds the --touch flag
which doesn't extract the file modified date and by that eliminates the
warnings.

Change-Id: I97e937aac233958db52b2fda685f5a5964fda2c8
Signed-off-by: Liron Aravot <laravot@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55163
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI