storage: Clean up image links during teardownImage
In commit 736ab9 we began cleaning up volume links that were created in
/var/run/vdsm/images/... Unfortunately on block domains we are leaving
around a broken symlink in /rhev/data-center/... in some cases and this
will cause storage operations related to a disk of a VM that was
previously running to fail on this host.  The solution is to also remove
this broken symlink during teardownImage.

Change-Id: I5fb871f3ccd785c555c72db34ab818c04d000534
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1319987
Signed-off-by: Adam Litke <alitke@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55164
Reviewed-by: Nir Soffer <nsoffer@redhat.com>
Continuous-Integration: Jenkins CI
Reviewed-by: Francesco Romani <fromani@redhat.com>