vm: handle destroy request while starting up
Since long time Vdsm allows Vm.destroy() requests
against VMs which are not yet fully created.
The most frequent use case is if the startup takes
lots of time, often because slow storage, so the
user decides to abort the startup.

Vdsm does the bare minimum to handle this flow,
and often misreports the Vm as down for error,
while it is not - the flow is supported.

To improve this situation, we add special explicit
handling for this case, and we add a new exitreason
to let Engine know that we had a destroy on startup.

Change-Id: I618eb03783d7059ae33d7b0a02542b99c8d8199b
Bug-Url: https://bugzilla.redhat.com/912390
Signed-off-by: Francesco Romani <fromani@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55151
Continuous-Integration: Jenkins CI
Reviewed-by: Dan Kenigsberg <danken@redhat.com>