image: add a flag for preparing with illegal volumes.
Currently preparing an image is impossible if one of its volumes is
illegal.

In the new image upload flow, the volume's legality is being used for
preventing volume usages while doing I/O using ovirt-imageio-daemon.

When the upload starts, a new image is created and prepared. Then, it is
being set to ILLEGAL, and only after- a ticket for doing I/O is being
sent to ovirt-imageio-daemon in order for the user to write data into
it.

If this process is stopped by any reason- resuming it means preparing
the image while it is ILLEGAL.

This patch adds a flag in the API for enabling the prepare process even
if its volumes are ILLEGAL, for the upload image to use.

The flow uses this flag because when a volume is illegal, it can't be
attached to a running VM, nor a VM can't start with an illegal disk
attached. Also, adding a new volume state, such as UPLOADING or
UNSAFE has implications on storage version, and means adding
validations in many storage flows.

Change-Id: Ie18ab167635e6e099c3a12e03ea6620578c3103f
Signed-off-by: Amit Aviram <aaviram@redhat.com>
Reviewed-on: https://gerrit.ovirt.org/55199
Continuous-Integration: Jenkins CI
Reviewed-by: Nir Soffer <nsoffer@redhat.com>